+++ docker/Fpm.Dockerfile
@@ -1,18 +1,5 @@
-#FROM php:7.3-fpm
 
-#RUN apt-get update\
-#&& docker-php-ext-install pdo pdo_mysql
         FROM php:fpm
-#RUN printf "deb http://archive.debian.org/debian/ jessie main\ndeb-src http://archive.debian.org/debian/ jessie main\ndeb http://security.debian.org jessie/updates main\ndeb-src http://security.debian.org jessie/updates main" > /etc/apt/sources.list
-#RUN echo "deb [check-valid-until=no] http://archive.debian.org/debian jessie-backports main" > /etc/apt/sources.list.d/jessie-backports.list
-
-# As suggested by a user, for some people this line works instead of the first one. Use whichever works for your case
-# RUN echo "deb [check-valid-until=no] http://archive.debian.org/debian jessie main" > /etc/apt/sources.list.d/jessie.list
-
-
-#RUN sed -i '/deb http:\/\/deb.debian.org\/debian jessie-updates main/d' /etc/apt/sources.list
-
-#UN apt-get -o Acquire::Check-Valid-Until=false update
         RUN apt-get update && apt-get install -y\
          libargon2-dev\
 		libcurl4-openssl-dev \
@@ -52,9 +39,3 @@
         && docker-php-ext-install pdo pdo_pgsql pdo_mysql
 
         RUN curl -sS https://getcomposer.org/installer | php -- --install-dir=/usr/local/bin --filename=composer
-       # COPY    ./site/ /var/www/html/
-        #COPY    ./config-php-fpm/php-fpm.conf  /usr/local/etc/php-fpm.conf
-       # COPY    ./www/www.conf   /usr/local/etc/php-fpm.d/www.conf
-        #EXPOSE 80 160
-
-        #CMD ["php-fpm"]
