+++ site/composer.json
@@ -1,0 +1,7 @@
+{
+    "autoload": {
+        "psr-4": {
+            "App\\": "src/app"
+        }
+    }
+}
